# backend/hollow/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /api/forest:
        post:
            tags:
                - Forests
            description: Push a Leaf
            operationId: Forests_PushLeaf
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/forest.v1.PushLeafRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/forest.v1.PushLeafReply'
    /api/forest/all:
        get:
            tags:
                - Forests
            description: Get Leafs
            operationId: Forests_GetForest
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: pagesize
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/forest.v1.GetLeafsReply'
    /api/forest/comments/{id}:
        put:
            tags:
                - Forests
            description: Like Comment
            operationId: Forests_LikeComment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/forest.v1.LikeCommentRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/forest.v1.LikeCommentReply'
        delete:
            tags:
                - Forests
            description: Delete Comment
            operationId: Forests_DeleteComment
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/forest.v1.DeleteCommentReply'
    /api/forest/comments/{root}:
        get:
            tags:
                - Forests
            description: Get Comments
            operationId: Forests_GetComments
            parameters:
                - name: root
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: pagesize
                  in: query
                  schema:
                    type: integer
                    format: int64
                - name: father
                  in: query
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/forest.v1.GetCommentsReply'
        post:
            tags:
                - Forests
            description: Comment Leaf
            operationId: Forests_Comment
            parameters:
                - name: root
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/forest.v1.CommentLeafRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/forest.v1.CommentLeafRePly'
    /api/forest/{id}:
        get:
            tags:
                - Forests
            description: Get Leaf Detail
            operationId: Forests_GetLeafDetail
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/forest.v1.GetLeafDetailReply'
        delete:
            tags:
                - Forests
            description: Delete a Leaf
            operationId: Forests_DeleteLeaf
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: integer
                    format: int64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/forest.v1.DeleteLeafReply'
...
components:
    schemas:
        forest.v1.Comment:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                owner:
                    type: integer
                    format: int64
                root:
                    type: integer
                    format: int64
                father:
                    type: integer
                    format: int64
                createdAt:
                    type: integer
                    format: int64
                status:
                    type: integer
                    format: int64
                message:
                    type: string
                liked:
                    type: integer
                    format: int64
        forest.v1.CommentLeafRePly:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
        forest.v1.CommentLeafRequest:
            type: object
            properties:
                root:
                    type: integer
                    format: int64
                status:
                    type: integer
                    format: int64
                message:
                    type: string
                father:
                    type: integer
                    format: int64
        forest.v1.DeleteCommentReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
        forest.v1.DeleteLeafReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
        forest.v1.GetCommentsReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
                data:
                    $ref: '#/components/schemas/forest.v1.MultipleCommentReply'
        forest.v1.GetLeafDetailReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
                data:
                    $ref: '#/components/schemas/forest.v1.Leaf'
        forest.v1.GetLeafsReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
                data:
                    $ref: '#/components/schemas/forest.v1.MultipleLeafReply'
        forest.v1.Leaf:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                owner:
                    type: integer
                    format: int64
                message:
                    type: string
                createdAt:
                    type: integer
                    format: int64
                status:
                    type: integer
                    format: int64
                liked:
                    type: integer
                    format: int64
        forest.v1.LikeCommentReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
        forest.v1.LikeCommentRequest:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
        forest.v1.MultipleCommentReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/forest.v1.Comment'
                total:
                    type: integer
                    format: int64
        forest.v1.MultipleLeafReply:
            type: object
            properties:
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/forest.v1.Leaf'
                total:
                    type: integer
                    format: int64
        forest.v1.PushLeafReply:
            type: object
            properties:
                code:
                    type: integer
                    format: int64
                msg:
                    type: string
        forest.v1.PushLeafRequest:
            type: object
            properties:
                status:
                    type: integer
                    format: int64
                message:
                    type: string
...